Deleted Audit class. Added Audit fields and methods for Entities Added BaseEntity as a Super class for all entities. Tags - changed id to Long. TagServiceImpl - fixed maxResults validation.

+++ gift-certificate-data/src/main/java/com/epam/esm/giftcertificatemodule3/dao/TagDAO.java
@@ -8,7 +8,7 @@ import com.epam.esm.giftcertificatemodule3.entity.Tag;
  *
  * @author Aleh Yemelyanchyk
  */
-public interface TagDAO extends CrudDAO<Tag, Integer> {
+public interface TagDAO extends CrudDAO<Tag, Long> {
     /**
      * Returns the instance of the Tag type.
      *
